requirements.txt
autopep8==1.4.4
blinker==1.4
Click==7.0
Flask==1.1.1
Flask-DebugToolbar==0.10.1
Flask-DotEnv==0.1.2
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
pycodestyle==2.5.0
python-dotenv==1.0.0
Werkzeug==0.16.0